Move TextEditor interface
