+++ src/main/webapp/app/shared/monaco-editor/model/actions/adapter/text-editor.interface.ts
