+++ src/main/webapp/app/shared/monaco-editor/model/actions/communication/monaco-channel-reference.action.ts
@@ -6,7 +6,7 @@ import { MetisService } from 'app/shared/metis/metis.service';
 import { ChannelService } from 'app/shared/metis/conversations/channel.service';
 import { firstValueFrom } from 'rxjs';
 import { CompletionItemKind, Disposable, EditorRange } from 'app/shared/monaco-editor/model/actions/monaco-editor.util';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 /**
  * Action to insert a reference to a channel into the editor. Users that type a # will see a list of available channels to reference.
