+++ src/main/webapp/app/shared/monaco-editor/model/actions/grading-criteria/monaco-grading-instruction.action.ts
@@ -4,7 +4,7 @@ import { MonacoGradingScaleAction } from 'app/shared/monaco-editor/model/actions
 import { MonacoGradingDescriptionAction } from 'app/shared/monaco-editor/model/actions/grading-criteria/monaco-grading-description.action';
 import { MonacoGradingFeedbackAction } from './monaco-grading-feedback.action';
 import { MonacoGradingUsageCountAction } from 'app/shared/monaco-editor/model/actions/grading-criteria/monaco-grading-usage-count.action';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 export class MonacoGradingInstructionAction extends MonacoEditorDomainAction {
     static readonly ID = 'monaco-grading-instruction.action';
