+++ src/main/webapp/app/shared/monaco-editor/model/actions/grading-criteria/monaco-grading-usage-count.action.ts
@@ -1,5 +1,5 @@
 import { MonacoEditorDomainAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-domain-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 export class MonacoGradingUsageCountAction extends MonacoEditorDomainAction {
     static readonly ID = 'monaco-grading-usage-count.action';
