+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-code-block.action.ts
@@ -1,6 +1,6 @@
 import { faFileCode } from '@fortawesome/free-solid-svg-icons';
 import { MonacoEditorAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 const CODE_BLOCK_DELIMITER = '```';
