+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-code.action.ts
@@ -1,6 +1,6 @@
 import { faCode } from '@fortawesome/free-solid-svg-icons';
 import { MonacoEditorAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 const CODE_DELIMITER = '`';
