+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-color.action.ts
@@ -1,5 +1,5 @@
 import { MonacoEditorAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 interface ColorArguments {
     color: string;
