+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-formula.action.ts
@@ -1,5 +1,5 @@
 import { MonacoEditorDomainAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-domain-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 const FORMULA_OPEN_DELIMITER = '$$ ';
 const FORMULA_CLOSE_DELIMITER = ' $$';
