+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-fullscreen.action.ts
@@ -1,7 +1,7 @@
 import { MonacoEditorAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-action.model';
 
 import { faCompress } from '@fortawesome/free-solid-svg-icons';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 /**
  * Action to toggle fullscreen mode in the editor.
