+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-test-case.action.ts
@@ -2,7 +2,7 @@ import { TranslateService } from '@ngx-translate/core';
 import { DomainActionWithOptionsArguments, MonacoEditorDomainActionWithOptions } from 'app/shared/monaco-editor/model/actions/monaco-editor-domain-action-with-options.model';
 import { CompletionItemKind, Disposable } from 'app/shared/monaco-editor/model/actions/monaco-editor.util';
 import { ValueItem } from 'app/shared/markdown-editor/value-item.model';
-import { TextEditor } from './adapter/text-editor-adapter.model';
+import { TextEditor } from './adapter/text-editor.interface';
 
 /**
  * Action to insert a test case into the editor. It also registers a completion item provider offers all possible test cases as completion items to the user.
