+++ src/main/webapp/app/shared/monaco-editor/model/actions/monaco-unordered-list.action.ts
@@ -1,7 +1,7 @@
 import { MonacoEditorAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-action.model';
 import { faListUl } from '@fortawesome/free-solid-svg-icons';
 import { makeEditorPosition, makeEditorRange } from 'app/shared/monaco-editor/model/actions/monaco-editor.util';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 const LIST_BULLET = '- ';
