+++ src/main/webapp/app/shared/monaco-editor/model/actions/quiz/monaco-wrong-multiple-choice-answer.action.ts
@@ -1,5 +1,5 @@
 import { MonacoEditorDomainAction } from 'app/shared/monaco-editor/model/actions/monaco-editor-domain-action.model';
-import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor-adapter.model';
+import { TextEditor } from 'app/shared/monaco-editor/model/actions/adapter/text-editor.interface';
 
 export class MonacoWrongMultipleChoiceAnswerAction extends MonacoEditorDomainAction {
     static readonly ID = 'monaco-incorrect-multiple-choice-answer.action';
